Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float stupeflip;
-    scanf("%f",&stupeflip);
+    double stupeflip;
+    scanf("%lf",&stupeflip);
     printf("You entered: %.4f",stupeflip);
     return 0;
 }
